Servicegraph and spanmetrics together.
TODO: connect operations
It may be worth considering adding labels to the source and target of a span in the OpenTelemetry Collector, as this would allow us to more accurately determine the relationships between different operations. This information is not currently available in metrics data but is present in trace storage. While metrics data can be a pragmatic overall approach, using a graph database may be a more effective way of understanding the relationships between different operations. Graph databases are currently receiving a lot of attention and discussion as a potential solution.
TODO: split spankind
Under cluster with yet another subgraph as client/server